feat: refactor structure for datetimeinput
