+++ frontend/src/widgets/inputs/DateTimeInputWidget/DateVariant.tsx
@@ -1,0 +1,104 @@
+import * as React from 'react';
+import { useState, useCallback } from 'react';
+import { Button } from '@/components/ui/button';
+import { Calendar } from '@/components/ui/calendar';
+import {
+  Popover,
+  PopoverContent,
+  PopoverTrigger,
+} from '@/components/ui/popover';
+import { format } from 'date-fns';
+import { Calendar as CalendarIcon } from 'lucide-react';
+import { cn } from '@/lib/utils';
+import { inputStyles } from '@/lib/styles';
+import { Scales } from '@/types/scale';
+import {
+  dateTimeInputVariants,
+  dateTimeInputIconVariants,
+  dateTimeInputTextVariants,
+} from '@/components/ui/input/date-time-input-variants';
+import { DateVariantProps } from './types';
+import { ActionIcons } from './shared';
+
+export const DateVariant: React.FC<DateVariantProps> = ({
+  value,
+  placeholder,
+  disabled,
+  nullable,
+  invalid,
+  onDateChange,
+  format: formatProp,
+  scale = Scales.Medium,
+  'data-testid': dataTestId,
+}) => {
+  const [open, setOpen] = useState(false);
+  const date = value ? new Date(value) : undefined;
+  const showClear = nullable && !disabled && value != null && value !== '';
+
+  const handleClear = (e?: React.MouseEvent) => {
+    e?.preventDefault();
+    e?.stopPropagation();
+    onDateChange(undefined);
+  };
+
+  const handleSelect = useCallback(
+    (selectedDate: Date | undefined) => {
+      onDateChange(selectedDate);
+      setOpen(false);
+    },
+    [onDateChange]
+  );
+
+  return (
+    <div className="relative w-full select-none">
+      <Popover open={open} onOpenChange={setOpen}>
+        <PopoverTrigger asChild>
+          <Button
+            disabled={disabled}
+            variant="outline"
+            className={cn(
+              dateTimeInputVariants({ scale }),
+              invalid && inputStyles.invalidInput,
+              disabled && 'cursor-not-allowed',
+              showClear && invalid
+                ? 'pr-16'
+                : showClear || invalid
+                  ? 'pr-8'
+                  : ''
+            )}
+            data-testid={dataTestId}
+          >
+            <CalendarIcon
+              className={cn(
+                'mr-2 shrink-0',
+                dateTimeInputIconVariants({ scale })
+              )}
+            />
+            <span
+              className={cn('truncate', dateTimeInputTextVariants({ scale }))}
+            >
+              {date
+                ? format(date, formatProp || 'yyyy-MM-dd')
+                : placeholder || 'Pick a date'}
+            </span>
+          </Button>
+        </PopoverTrigger>
+        <PopoverContent className="w-auto p-0" align="start">
+          <Calendar
+            mode="single"
+            selected={date}
+            onSelect={handleSelect}
+            initialFocus
+            scale={scale}
+          />
+        </PopoverContent>
+      </Popover>
+      <ActionIcons
+        showClear={showClear}
+        invalid={invalid}
+        onClear={handleClear}
+        scale={scale}
+      />
+    </div>
+  );
+};
