+++ frontend/src/widgets/inputs/DateTimeInputWidget/TimeVariant.tsx
@@ -1,0 +1,139 @@
+import * as React from 'react';
+import { useState, useCallback } from 'react';
+import { Input } from '@/components/ui/input';
+import { format } from 'date-fns';
+import { Clock } from 'lucide-react';
+import { cn } from '@/lib/utils';
+import { inputStyles } from '@/lib/styles';
+import { Scales } from '@/types/scale';
+import {
+  dateTimeInputIconVariants,
+  dateTimeInputTextVariants,
+} from '@/components/ui/input/date-time-input-variants';
+import { TimeVariantProps } from './types';
+import { ActionIcons } from './shared';
+
+export const TimeVariant: React.FC<TimeVariantProps> = ({
+  value,
+  placeholder,
+  disabled,
+  nullable,
+  invalid,
+  onTimeChange,
+  scale = Scales.Medium,
+  'data-testid': dataTestId,
+}) => {
+  // Use local state for the input value to make it uncontrolled
+  const [localTimeValue, setLocalTimeValue] = useState(() => {
+    if (value) {
+      // Parse the value to get time in HH:mm:ss format
+      try {
+        const date = new Date(value);
+        if (!isNaN(date.getTime())) {
+          return format(date, 'HH:mm:ss');
+        }
+      } catch {
+        // If parsing fails, try to use the value directly if it looks like a time
+        if (
+          typeof value === 'string' &&
+          /^\d{1,2}:\d{2}(:\d{2})?$/.test(value)
+        ) {
+          return value.length <= 5 ? value + ':00' : value;
+        }
+      }
+    }
+    return '00:00:00';
+  });
+
+  // Update local state when value prop changes (from parent)
+  React.useEffect(() => {
+    if (value) {
+      try {
+        const date = new Date(value);
+        if (!isNaN(date.getTime())) {
+          const newTimeValue = format(date, 'HH:mm:ss');
+          setLocalTimeValue(newTimeValue);
+        }
+      } catch {
+        // If parsing fails, try to use the value directly if it looks like a time
+        if (
+          typeof value === 'string' &&
+          /^\d{1,2}:\d{2}(:\d{2})?$/.test(value)
+        ) {
+          const newTimeValue = value.length <= 5 ? value + ':00' : value;
+          setLocalTimeValue(newTimeValue);
+        }
+      }
+    } else {
+      setLocalTimeValue('00:00:00');
+    }
+  }, [value]);
+
+  const showClear = nullable && !disabled && value != null && value !== '';
+
+  const handleClear = (e?: React.MouseEvent) => {
+    e?.preventDefault();
+    e?.stopPropagation();
+    onTimeChange('');
+  };
+
+  const handleTimeChange = useCallback(
+    (e: React.ChangeEvent<HTMLInputElement>) => {
+      const newTimeValue = e.target.value;
+      setLocalTimeValue(newTimeValue);
+    },
+    []
+  );
+
+  const handleTimeBlur = useCallback(() => {
+    // When input loses focus, update the parent
+    onTimeChange(localTimeValue);
+  }, [localTimeValue, onTimeChange]);
+
+  const handleKeyDown = useCallback(
+    (e: React.KeyboardEvent<HTMLInputElement>) => {
+      // When user presses Enter, update the parent
+      if (e.key === 'Enter') {
+        onTimeChange(localTimeValue);
+      }
+    },
+    [localTimeValue, onTimeChange]
+  );
+
+  return (
+    <div className="relative w-full select-none" data-testid={dataTestId}>
+      <div className="relative">
+        <Clock
+          className={cn(
+            'absolute left-3 top-1/2 -translate-y-1/2 shrink-0',
+            dateTimeInputIconVariants({ scale }),
+            disabled && 'opacity-50'
+          )}
+        />
+        <Input
+          type="time"
+          step="1"
+          value={localTimeValue}
+          onChange={handleTimeChange}
+          onBlur={handleTimeBlur}
+          onKeyDown={handleKeyDown}
+          disabled={disabled}
+          placeholder={placeholder || 'Select time'}
+          className={cn(
+            'bg-transparent appearance-none [&::-webkit-calendar-picker-indicator]:hidden cursor-pointer w-full pl-10',
+            dateTimeInputTextVariants({ scale }),
+            invalid && inputStyles.invalidInput,
+            disabled && 'cursor-not-allowed',
+            showClear && invalid ? 'pr-16' : showClear || invalid ? 'pr-8' : ''
+          )}
+        />
+      </div>
+      <ActionIcons
+        showClear={showClear}
+        invalid={invalid}
+        onClear={handleClear}
+        scale={scale}
+      />
+    </div>
+  );
+};
