+++ frontend/src/widgets/inputs/DateTimeInputWidget/index.ts
@@ -1,0 +1,9 @@
+export { DateTimeInputWidget } from './DateTimeInputWidget';
+export type {
+  DateTimeInputWidgetProps,
+  VariantType,
+  DateVariantProps,
+  DateTimeVariantProps,
+  TimeVariantProps,
+  BaseVariantProps,
+} from './types';
